Refactor code to fit with rest of naming convention. Add proper fetching into the page. Remove old test code.
